Refactor show locals.
